combained-tableinspector-into-fileuploader
